Fix selection segfault (#136)
* fix segfault when user attempts to move bits

When a user attempts to select and move both bits and lines, the program segfaults. This occurs when the bit objects that are referenced by the selection are being deleted. This solution resolves the issue by first removing the bits from the selection before deleting them.

* remove extra call to clearBits

There is no need for clearBits to be called when the right mouse button is pressed. The only action tied with the right mouse button that requires a call to clearBits is moveLines, which calls clearBits already.